schema: f64->xsd:double, bigdecimal xsd:decimal field type, strict v12 reads
Completes the M1 numeric overhaul for TerminusDB 12 (arbitrary_precision landed
separately):

- f64 now maps to xsd:double (IEEE binary64); f32 stays xsd:float. Was
  incorrectly xsd:float for both.
- New first-class xsd:decimal field type via bigdecimal::BigDecimal (arbitrary
  precision — v12 stores up to 256 digits). Serialized string-wrapped on write
  (exact, no float coercion; TerminusDB accepts string decimal input), read from
  native number or string. Re-exported as terminusdb_schema::BigDecimal so models
  need no direct bigdecimal dep. Dropped the unused rust_decimal dependency.
- Removed the v11 string-wrapped-number read fallbacks (strict v12: numbers are
  native JSON numbers).
- Replaced serde_canonical_json in json/normalization.rs with a hand-written
  arbitrary_precision-safe canonical JSON serializer (the crate panicked on the
  arbitrary_precision number token, breaking client-side content hashing).

Verified: schema unit tests green (incl. 20-digit decimal round-trip);
client integration suite 248/248 green on 12.1; and a live end-to-end test
inserts a 25-significant-digit decimal field and reads it back exactly.

+++ crates/schema/src/primitive.rs
@@ -184,7 +184,8 @@ to_schema_class!({
     i64: LONG,
     isize: INTEGER,
     String: STRING,
-    f64: FLOAT,
+    // f64 is IEEE-754 binary64 == xsd:double; f32 is binary32 == xsd:float.
+    f64: DOUBLE,
     f32: FLOAT,
     bool: BOOLEAN,
     u8: UNSIGNED_BYTE,
